factor out glyph change funcs into dedicated module
